Basic form fields tests
